Implement sampling for MTP
